+++ crypto3/libs/zk/include/nil/crypto3/zk/snark/arithmetization/plonk/lookup_gate.hpp
@@ -49,11 +49,11 @@ namespace nil {
                     plonk_lookup_gate(  ) : tag_index(0) {
                     }
 
-                    plonk_lookup_gate( std::size_t tag_index, const ConstraintType &constraint ) :
+                    plonk_lookup_gate( std::int64_t tag_index, const ConstraintType &constraint ) :
                         constraints(std::vector<ConstraintType>(1, constraint)), tag_index(tag_index) {
                     }
 
-                    plonk_lookup_gate( std::size_t tag_index, const std::vector<ConstraintType> &constraints) :
+                    plonk_lookup_gate( std::int64_t tag_index, const std::vector<ConstraintType> &constraints) :
                         constraints(constraints), tag_index(tag_index) {
                     }
