+++ crypto3/libs/zk/include/nil/crypto3/zk/snark/arithmetization/plonk/lookup_table.hpp
@@ -56,7 +56,7 @@ namespace nil {
                     plonk_lookup_table() : columns_number(0), tag_index(0) {
                     }
 
-                    plonk_lookup_table(std::size_t _columns_number, std::size_t _tag_index) :
+                    plonk_lookup_table(std::size_t _columns_number, std::int64_t _tag_index) :
                         columns_number(_columns_number), tag_index(_tag_index) {
                     }
